open_ai_agents/pyproject.toml
[project]
name = "open-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "black>=25.12.0",
    "cachetools>=6.2.4",
    "dotenv>=0.9.9",
    "openai-agents>=0.6.7",
    "ruff>=0.14.13",
    "sendgrid>=6.12.5",
]
